+++ packages/lesswrong/themes/forumTheme.ts
@@ -1,7 +1,7 @@
 import { getForumType, ThemeOptions } from './themeNames';
 import { baseTheme } from './createThemeDefaults';
 import { getUserTheme } from './userThemes/index';
-import { getSiteTheme } from './siteThemes/index';
+import { getSiteTheme } from './siteThemes';
 import type { ForumTypeString } from '../lib/instanceSettings';
 import deepmerge from 'deepmerge';
 import { forumSelect } from '../lib/forumTypeUtils';
