+++ packages/lesswrong/themes/siteThemes.ts
@@ -1,0 +1,23 @@
+import { ForumTypeString } from '@/lib/instanceSettings';
+
+export const lessWrongTheme: SiteThemeSpecification = {
+  componentPalette: (dark: boolean) => ({
+  }),
+  make: (palette: ThemePalette) => ({
+    isLW: true,
+    isAF: false,
+  }),
+};
+export const alignmentForumTheme: SiteThemeSpecification = {
+  componentPalette: (dark: boolean) => ({
+  }),
+  make: (palette: ThemePalette) => ({
+    isLW: true,
+    isAF: false,
+  }),
+};
+
+export const getSiteTheme = (forumType: ForumTypeString): SiteThemeSpecification => {
+  if (forumType === 'AlignmentForum') return alignmentForumTheme;
+  else return lessWrongTheme;
+}
