--- packages/lesswrong/themes/siteThemes/alignmentForumTheme.ts
@@ -1,11 +1,0 @@
-export const alignmentForumTheme: SiteThemeSpecification = {
-  componentPalette: (dark: boolean) => ({
-  }),
-  make: (palette: ThemePalette) => ({
-    isLW: true,
-    isAF: false,
-  }),
-};
-
-
-export default alignmentForumTheme;
