--- packages/lesswrong/themes/siteThemes/lesswrongTheme.ts
@@ -1,8 +1,0 @@
-export const lessWrongTheme: SiteThemeSpecification = {
-  componentPalette: (dark: boolean) => ({
-  }),
-  make: (palette: ThemePalette) => ({
-    isLW: true,
-    isAF: false,
-  }),
-};
